Keep the same project-folder structure
